[BE] test : 모임 삭제 테스트

+++ backend/src/main/java/moim_today/implement/meeting/joined_meeting/JoinedMeetingRemover.java
@@ -17,6 +17,8 @@ public class JoinedMeetingRemover {
 
     @Transactional
     public void deleteAllByMeetingIdIn(final List<Long> meetingIds) {
-        joinedMeetingRepository.deleteAllByMeetingIdIn(meetingIds);
+        if (!meetingIds.isEmpty()) {
+            joinedMeetingRepository.deleteAllByMeetingIdIn(meetingIds);
+        }
     }
 }
